requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
bcrypt==4.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
lxml==4.9.1
psycopg2-binary==2.9.5
PyJWT==2.6.0
python-dotenv==0.21.0
pytz==2022.7
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.13
